[policies] Fix dist_version for Ubuntu
DISTRIB_RELEASE variable contains a float value for the Ubuntu
release version and not an integer value.

Example, a 22.04 release is an LTS release, and a 22.10 release
is an interim non-LTS release, changing this value to a float
allows us to retrieve the full version rather than just the int.

+++ sos/policies/distros/ubuntu.py
@@ -70,7 +70,7 @@ class UbuntuPolicy(DebianPolicy):
                 lines = fp.readlines()
                 for line in lines:
                     if "DISTRIB_RELEASE" in line:
-                        return int(line.split("=")[1].strip())
+                        return float(line.split("=")[1].strip())
             return False
         except (IOError, ValueError):
             return False
